Implementado os pinguins e algumas funções em Vec2 e Rect
Roteiro 6 etapa 2.

+++ include/Vec2.h
@@ -8,7 +8,7 @@
 		float x, y;
 
 		Vec2(){}
-		Vec2(int, int);
+		Vec2(float, float);
 
 		void Rotate(float);
 		float Modulo();
