+++ src/Vec2.cpp
@@ -1,6 +1,6 @@
 #include "Vec2.h"
 
-Vec2::Vec2(int x, int y) {
+Vec2::Vec2(float x, float y) {
 	this->x = x;
 	this->y = y;
 }
